+++ LICENSE
@@ -1,6 +1,7 @@
 This software is licensed under the Apache 2 license, quoted below.
 
 Copyright 2016-2017 Azavea [http://www.azavea.com]
+Copyright 2018 CCRi [http://www.ccri.com]
 
 Licensed under the Apache License, Version 2.0 (the "License"); you may not
 use this file except in compliance with the License. You may obtain a copy of
