+++ geopyspark-backend/build.sbt
@@ -66,4 +66,9 @@ lazy val geotoolsProject = Project("geotools", file("geotools"))
 lazy val geotrellisProject = Project("geotrellis-backend", file("geotrellis"))
   .settings(commonSettings: _*)
   .settings(publishSettings: _*)
-  .dependsOn(utilProject, vectorPipeProject, geotoolsProject)
+  .dependsOn(utilProject, vectorPipeProject)
+
+lazy val geopysparkProject = Project("geopyspark", file("geopyspark"))
+  .settings(commonSettings: _*)
+  .settings(publishSettings: _*)
+  .dependsOn(utilProject, vectorPipeProject, geotoolsProject, geotrellisProject)
