+++ geopyspark-backend/geopyspark/build.sbt
@@ -1,0 +1,20 @@
+name := "geopyspark"
+
+libraryDependencies ++= Seq(
+  "org.locationtech.geomesa" %% "geomesa-spark-core" % Version.geomesa
+)
+
+assemblyMergeStrategy in assembly := {
+  case "reference.conf" => MergeStrategy.concat
+  case "application.conf" => MergeStrategy.concat
+  case "META-INF/MANIFEST.MF" => MergeStrategy.discard
+  case "META-INF\\MANIFEST.MF" => MergeStrategy.discard
+  case "META-INF/ECLIPSEF.RSA" => MergeStrategy.discard
+  case "META-INF/ECLIPSEF.SF" => MergeStrategy.discard
+  case x if x.startsWith("META-INF/services") => MergeStrategy.concat
+  case _ => MergeStrategy.first
+}
+
+PB.targets in Compile := Seq(
+  scalapb.gen() -> (sourceManaged in Compile).value
+)
